docs(PhoneInput): fix story title
